fix: rename and move to HOC folder
