+++ src/index.d.ts
@@ -93,4 +93,4 @@ export { default as VisualPickerOption } from './components/VisualPickerOption';
 export { default as VisualPickerOptionFooter } from './components/VisualPickerOptionFooter';
 export { default as WeekDayPicker } from './components/WeekDayPicker';
 export { default as WeeklyCalendar } from './components/WeeklyCalendar';
-export { default as RainbowTheme } from './libs/RainbowTheme';
+export { default as withRainbowTheme } from './libs/hocs/withRainbowTheme';
