+++ src/index.js
@@ -92,4 +92,4 @@ export { default as VisualPickerOptionFooter } from './components/VisualPickerOp
 export { default as WeekDayPicker } from './components/WeekDayPicker';
 export { default as WeeklyCalendar } from './components/WeeklyCalendar';
 export { default as withDebounce } from './libs/hocs/withDebounce';
-export { default as RainbowTheme } from './libs/RainbowTheme';
+export { default as withRainbowTheme } from './libs/hocs/withRainbowTheme';
