+++ src/libs/hocs/withRainbowTheme/index.js
@@ -1,11 +1,11 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import React from 'react';
-import RainbowThemeContainer from '../../components/RainbowThemeContainer';
+import RainbowThemeContainer from '../../../components/RainbowThemeContainer';
 
-const RainbowTheme = (Component, theme) => props => (
+const withRainbowTheme = (Component, theme) => props => (
     <RainbowThemeContainer theme={theme}>
         <Component {...props} />
     </RainbowThemeContainer>
 );
 
-export default RainbowTheme;
+export default withRainbowTheme;
